fix(./tasks/*): Some fixes after restoring tasks of other persons
