+++ tasks/remizov_k_max_in_matrix_string/info.json
@@ -1,9 +1,9 @@
 {
   "student": {
     "first_name": "Кирилл",
-    "last_name": "Львович",
-    "middle_name": "Ремизов",
+    "last_name": "Ремизов",
+    "middle_name": "Львович",
     "group_number": "3823Б1ПР4",
-    "task_number": "3"
+    "task_number": "1"
   }
 }
