+++ tasks/remizov_k_systems_linear_equations_gradient/info.json
@@ -1,0 +1,9 @@
+{
+  "student": {
+    "first_name": "Кирилл",
+    "last_name": "Львович",
+    "middle_name": "Ремизов",
+    "group_number": "3823Б1ПР4",
+    "task_number": "3"
+  }
+}
