+++ tasks/remizov_k_systems_linear_equations_gradient/settings.json
@@ -1,7 +1,7 @@
 {
   "tasks_type": "processes",
   "tasks": {
-    "mpi": "disabled",
-    "seq": "disabled"
+    "mpi": "enabled",
+    "seq": "enabled"
   }
 }
